+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/index.ts
@@ -7,4 +7,4 @@ export * from './table-menu/o-table-export-configuration.class';
 export * from './table-menu/o-table-menu.component';
 export * from './table-option/o-table-option.component';
 export * from './table-quickfilter/o-table-quickfilter.component';
-export * from './table-columns-filter/columns/o-table-columns-filter-columns.component';
+export * from './table-columns-filter/columns/o-table-columns-filter-column.component';
