+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/o-table-header-components.ts
@@ -6,7 +6,7 @@ import { OTableInsertableRowComponent } from './table-insertable-row/o-table-ins
 import { OTableMenuComponent } from './table-menu/o-table-menu.component';
 import { OTableOptionComponent } from './table-option/o-table-option.component';
 import { OTableQuickfilterComponent } from './table-quickfilter/o-table-quickfilter.component';
-import { OTableColumnsFilterColumnComponent } from './table-columns-filter/columns/o-table-columns-filter-columns.component';
+import { OTableColumnsFilterColumnComponent } from './table-columns-filter/columns/o-table-columns-filter-column.component';
 
 export const O_TABLE_HEADER_COMPONENTS = [
   OTableButtonComponent,
