+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/table-columns-filter/o-table-columns-filter.component.ts
@@ -5,7 +5,7 @@ import { Codes } from '../../../../../util/codes';
 import { Util } from '../../../../../util/util';
 import { OColumn } from '../../../column/o-column.class';
 import { OTableComponent } from '../../../o-table.component';
-import { OTableColumnsFilterColumnComponent } from './columns/o-table-columns-filter-columns.component';
+import { OTableColumnsFilterColumnComponent } from './columns/o-table-columns-filter-column.component';
 
 export const DEFAULT_INPUTS_O_TABLE_COLUMN_FILTER = [
   // columns [string]: columns that might be filtered, separated by ';'. Default: all visible columns.
